+++ scorex-basics/src/main/scala/scorex/transaction/StateChangeReason.scala
@@ -3,7 +3,7 @@ package scorex.transaction
 /**
   * reason to change account balance
   */
-trait StateChangeReason {
+trait StateChangeReason extends Serializable {
   /**
     * A transaction could be serialized into binary form
     */