+++ release-notes.md
@@ -1,7 +1,10 @@
 **1.2.4**
 
 * New API call /transactions/info
+
 * MVStore is used for state persistence
+* Fixed JSON style in API calls
+* PeersHttpService was rebaned to PeersApiRoute
 
 **1.2.3**